+++ webui/webpack.config.js
@@ -1,26 +1,33 @@
 const { join } = require('path');
 
 const { createWebpackConfiguration, initEnv } = require('@tarantool.io/webpack-config');
+
 const { namespace } = require('./module-config');
 
 const env = initEnv();
 
-const root = __dirname;
-const entry = process.env.WEBPACK_APP_ENTRY || 'index.js';
+const isServe = process.env['WEBPACK_SERVE'] === 'true';
+const isProd = env['NODE_ENV'] === 'production';
 
-const isProd = env.NODE_ENV === 'production';
+const root = __dirname;
+const entry = join(root, 'src', process.env['WEBPACK_APP_ENTRY'] || 'index.js');
+const build = join(root, 'build');
+const htmlTemplate = isServe ? join(root, 'public', 'index.html') : undefined;
+const proxy = process.env.WEBPACK_DEV_SERVER_PROXY === 'true' ? require('./config/proxy.config') : undefined;
+const sourceMap = process.env.GENERATE_SOURCEMAP === 'true';
 
 module.exports = createWebpackConfiguration({
   namespace,
   root,
-  entry: join(root, 'src', entry),
-  htmlTemplate: !isProd ? join(root, 'public', 'index.html') : undefined,
-  lua: true,
+  entry,
+  build,
+  htmlTemplate,
+  lua: isProd && !isServe,
   lint: true,
   emotion: true,
   env,
-  sourceMap: process.env.GENERATE_SOURCEMAP === 'true',
-  proxy: process.env.WEBPACK_DEV_SERVER_PROXY === 'true' ? require('./config/proxy.config') : undefined,
+  sourceMap,
+  proxy,
   externals: isProd
     ? {
         react: 'react',